remove utils file on ChartMenu
